backend/requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
langchain
python-magic
pydantic-settings
pypdf
google-genai
requests
langchain
langchain-google-genai
langchain-text-splitters